change NoLoginHomeGallery component name
