+++ src/components/home/HomeGallery.tsx
@@ -1,65 +1,77 @@
-// "use client";
+"use client";
 
-// import React, { ReactNode, useEffect, useState } from "react";
-// import { useInView } from "react-intersection-observer";
-// import { fetchMoreLatestPostsComponent } from "@/lib/fetchWrapper";
+import React, { useEffect, useState } from "react";
+import ImageFront from "./ImageFront";
+import ImageBack from "./ImageBack";
+import { GalleyPost, UserRelationship } from "@/lib/definitions";
+import { fetchMoreLatestPosts } from "@/lib/fetch";
+import { useInView } from "react-intersection-observer";
+import { useCursorById } from "@/lib/utils";
+import HomeFlipImage from "./HomeFlipImage";
 
-// type Props = {
-//   firstPosts: ReactNode[][];
-//   cursorId: string;
-// };
+type Props = {
+  firstPosts: GalleyPost[][];
+  myId: string | undefined;
+  cursorId: string;
+};
 
-// const HomeGallery = (props: Props) => {
-//   const { firstPosts, cursorId } = props;
-//   //   const [posts, setPosts] = useState<ReactNode[][]>(firstPosts);
+const HomeGallery = (props: Props) => {
+  const { firstPosts, myId, cursorId } = props;
+  const { cursorById } = useCursorById();
+  const [posts, setPosts] = useState<GalleyPost[][]>(firstPosts);
+  const [cursorPostId, setCursorPostId] = useState(cursorId);
+  const [postLimit, setPostLimit] = useState(false);
+  const { ref, inView } = useInView({
+    threshold: 0,
+    initialInView: undefined,
+  });
 
-//   //   const [cursorPostId, setCursorPostId] = useState(cursorId);
-//   //   const [postLimit, setPostLimit] = useState(false);
-//   //   const { ref, inView } = useInView({
-//   //     threshold: 0,
-//   //     initialInView: undefined,
-//   //   });
+  useEffect(() => {
+    if (inView && !postLimit) {
+      const fetchMorePosts = async () => {
+        const newPosts = await fetchMoreLatestPosts(6, null, cursorPostId);
+        if (newPosts.length < 6) {
+          setPostLimit(true);
+        }
+        const newPostsArray: GalleyPost[][] = [[], [], []];
+        newPosts.forEach((post: GalleyPost, i) => {
+          newPostsArray[i % 3] = [...newPostsArray[i % 3], post];
+        });
 
-//   //   useEffect(() => {
-//   //     if (inView && !postLimit) {
-//   //       const fetchMorePosts = async () => {
-//   //         const { component, cursorId } = await fetchMoreLatestPostsComponent(
-//   //           6,
-//   //           null,
-//   //           cursorPostId
-//   //         );
-//   //         if (component.length < 6) {
-//   //           setPostLimit(true);
-//   //         }
-//   //         const newPostsArray: ReactNode[][] = [[], [], []];
-//   //         component.forEach((post, i) => {
-//   //           newPostsArray[i % 3] = [...newPostsArray[i % 3], post];
-//   //         });
+        setPosts((prevPosts) => [
+          [...prevPosts[0], ...newPostsArray[0]],
+          [...prevPosts[1], ...newPostsArray[1]],
+          [...prevPosts[2], ...newPostsArray[2]],
+        ]);
+        const newCursorPostId = cursorById(newPosts);
+        setCursorPostId(newCursorPostId);
+      };
+      fetchMorePosts();
+    }
+  }, [inView]);
 
-//   //         setPosts((prevPosts) => [
-//   //           [...prevPosts[0], newPostsArray[0]],
-//   //           [...prevPosts[1], newPostsArray[1]],
-//   //           [...prevPosts[2], newPostsArray[2]],
-//   //         ]);
-//   //         setCursorPostId(cursorId);
-//   //       };
-//   //       fetchMorePosts();
-//   //     }
-//   //   }, [inView]);
-//   //   return (
-//   //     <div className="lg:px-40 px-5 flex">
-//   //       {posts.map((colPosts, col) => (
-//   //         <div key={col} className="relative w-1/3 p-1 lg:p-2">
-//   //           {colPosts}
-//   //           <div ref={ref} className="absolute bottom-[500px] h-[1px]" />
-//   //         </div>
-//   //       ))}
-//   //     </div>
-//   return false ? (
-//     <LoginHomeGallery firstPost={posts} />
-//   ) : (
-//     <NoLoginHomeGallery firstPost={posts} myId={session?.user.id} />
-//   );
-// };
+  return (
+    <div className="lg:px-40 px-5 flex">
+      {posts.map((colPosts: GalleyPost[]) => (
+        <div key={colPosts[0].id} className="w-1/3 p-1 lg:p-2">
+          {colPosts.map((post: GalleyPost, index) => (
+            <HomeFlipImage
+              key={post.id}
+              post={post}
+              myId={myId}
+              containerStyle={{
+                width: "100%",
+                height: "auto",
+                marginBottom: "8px",
+              }}
+              frontComponent={<ImageFront index={index} post={post} />}
+            />
+          ))}
+          <div ref={ref} />
+        </div>
+      ))}
+    </div>
+  );
+};
 
-// export default HomeGallery;
+export default HomeGallery;
