--- src/components/home/HomeGalleryFetch.tsx
@@ -1,16 +1,0 @@
-// import React from "react";
-// import HomeGallery from "./HomeGallery";
-// import { GalleyPost } from "@/lib/definitions";
-
-// type Props = {
-//   firstPosts: GalleyPost[][];
-//   myId: string | undefined;
-// };
-
-// const HomeGalleryFetch = async (props: Props) => {
-//   const { firstPosts, myId } = props;
-
-//   return <HomeGallery firstPosts={firstPosts} cursorId={cursorId} />;
-// };
-
-// export default HomeGalleryFetch;
